# meta/tikv: simplify transactions in point get scenarios by reducing one RPC call