--- a/pkg/meta/tkv_mem.go
+++ b/pkg/meta/tkv_mem.go
@@ -203,6 +203,10 @@ func (c *memKV) set(key string, value []byte) {
 	}
 }
 
+func (c *memKV) pointGetTxn(ctx context.Context, f func(*kvTxn) error, retry int) (err error) {
+	return c.txn(ctx, f, retry)
+}
+
 func (c *memKV) txn(ctx context.Context, f func(*kvTxn) error, retry int) error {
 	tx := &memTxn{
 		store:    c,