--- a/pkg/meta/tkv_prefix.go
+++ b/pkg/meta/tkv_prefix.go
@@ -80,8 +80,8 @@ type prefixClient struct {
 	prefix []byte
 }
 
-func (c *prefixClient) pointGetTxn(ctx context.Context, f func(*kvTxn) error, retry int) (err error) {
-	return c.tkvClient.pointGetTxn(ctx, func(tx *kvTxn) error {
+func (c *prefixClient) simpleTxn(ctx context.Context, f func(*kvTxn) error, retry int) (err error) {
+	return c.tkvClient.simpleTxn(ctx, func(tx *kvTxn) error {
 		return f(&kvTxn{&prefixTxn{tx, c.prefix}, retry})
 	}, retry)
 }