--- a/pkg/meta/tkv_tikv.go
+++ b/pkg/meta/tkv_tikv.go
@@ -24,7 +24,6 @@ import (
 	"math"
 	"net/url"
 	"os"
-	"runtime/debug"
 	"strings"
 	"syscall"
 	"time"
@@ -199,7 +198,7 @@ func (c *tikvClient) config(key string) interface{} {
 	return nil
 }
 
-func (c *tikvClient) pointGetTxn(ctx context.Context, f func(*kvTxn) error, retry int) (err error) {
+func (c *tikvClient) simpleTxn(ctx context.Context, f func(*kvTxn) error, retry int) (err error) {
 	tx, err := c.client.Begin(tikv.WithStartTS(math.MaxUint64)) // math.MaxUint64 means to point get the latest committed data without PD access
 	if err != nil {
 		return errors.Wrap(err, "failed to begin transaction")
@@ -218,7 +217,6 @@ func (c *tikvClient) pointGetTxn(ctx context.Context, f func(*kvTxn) error, retr
 		return err
 	}
 	if !tx.IsReadOnly() {
-		logger.Errorf("TiKV point get transaction is read-only\n%s", debug.Stack()) // should not happen
 		return syscall.EINVAL
 	}
 	return nil