In the current implementation of TiKV, each transaction requires at least two RPCs: one to obtain start timestamp from PD, and another to read/write data from TiKV, even for the simplest point query scenarios such as `lookup/getattr`. Actually, TiKV has optimized for simple point query scenarios by allowing use of `math.MaxUint64` as a start timestamp, which internally means "to get the latest committed data." This way, we can eliminate one RPC call to PD, reducing latency and alleviating the load on PD.

This optimization corresponds to [simpleTxn](https://github.com/juicedata/juicefs/pull/5377) in SQL engine - to reduce transaction overhead for simple queries. It provides a significant improvement for read-heavy scenarios, such as LLM training. A simple `mdtest` benchmark shows that `stat` operations can achieve a 50%-60% performance improvement after this optimization.

Before:
```
SUMMARY rate: (of 5 iterations)
   Operation                     Max            Min           Mean        Std Dev
   ---------                     ---            ---           ----        -------
   Directory creation            555.637        500.848        535.327         21.022
   Directory stat              30217.822      24352.151      26544.440       2232.511
   Directory rename              521.461        493.811        509.121          9.932
   Directory removal             519.550        487.976        506.651         11.587
   File creation                 679.972        640.534        655.575         14.717
   File stat                   30788.481      26540.856      29316.366       1654.854
   File read                   37628.849      27392.933      35024.104       4340.994
   File removal                  749.882        703.259        724.518         23.567
   Tree creation                 592.254        483.587        544.636         41.305
   Tree removal                  527.064        495.100        511.837         15.593
```

After:
```
SUMMARY rate: (of 5 iterations)
   Operation                     Max            Min           Mean        Std Dev
   ---------                     ---            ---           ----        -------
   Directory creation            802.445        751.059        777.577         19.036
   Directory stat              47103.933      39174.624      44248.699       3093.214
   Directory rename              552.234        527.294        537.524          9.463
   Directory removal             541.498        512.040        524.373         13.788
   File creation                 791.234        731.998        768.850         23.305
   File stat                   45019.831      41267.642      43179.429       1789.069
   File read                   35994.526      27491.580      32485.700       3678.509
   File removal                  784.541        679.333        741.993         41.675
   Tree creation                 794.911        725.299        775.506         28.566
   Tree removal                  649.409        581.852        628.249         27.925
```