@polyrabbit 2025-05-28
Update: #6150 should have fixed the flaky test - it seems the `RC` isolation level does not work well with `async_commit`, not sure if it's a bug in TiKV.