@lidaohang 2025-08-11
Before:
<img width="835" height="362" alt="image" src="https://github.com/user-attachments/assets/eb820826-7b71-4fbf-813f-ac1e5e4872ad" />


After:
<img width="842" height="362" alt="image" src="https://github.com/user-attachments/assets/75624fe2-0909-45dc-b679-4327e546a3bc" />